Add porting stages
The description of porting stages is not complete.
Many terms are not introduced in the previous sections and there
are some explaination about the system that must be described in
the architecture chapters. Moreover we need a better description
of the porting stages, the description in this commit is only
a summary.